프로모터 패키지 변경
