+++ src/main/java/net/admins/service/StaffService.java
@@ -30,25 +30,4 @@ public interface StaffService {
 	
 	int insertStfRegist(Map<String, Object> map);
 
-	int selectPmtOneGetTotalCount(CommandMap commandMap);
-
-	List<Map<String, Object>> selectListPmtList(Map<String, Object> map);
-
-	void insertPmtForm(CommandMap commandMap) throws Exception;
-
-	boolean selectOnePmtLoginIdCheck(String stfLoginId) throws Exception;
-
-	Map<String, Object> selectPmtView(String promoter_id);
-
-	Map<String, Object> updatePmtForm(String promoter_id);
-
-	void updatePmt(CommandMap commandMap);
-
-	boolean updatePmtPassword(CommandMap commandMap) throws IOException;
-
-	void leavePmt(CommandMap commandMap);
-
-	Map<String, Object> selectStaffSeek(CommandMap commandMap);
-
-	void returnPmt(CommandMap commandMap);
 }
